+++ R/RcppExports.R
@@ -17,6 +17,38 @@ EigenR_det_sparse_cplx <- function(i, j, Mij, nrows, ncols) {
     .Call('_EigenR_EigenR_det_sparse_cplx', PACKAGE = 'EigenR', i, j, Mij, nrows, ncols)
 }
 
+EigenR_absdet <- function(M) {
+    .Call('_EigenR_EigenR_absdet', PACKAGE = 'EigenR', M)
+}
+
+EigenR_logabsdet <- function(M) {
+    .Call('_EigenR_EigenR_logabsdet', PACKAGE = 'EigenR', M)
+}
+
+EigenR_isInjective_real <- function(M) {
+    .Call('_EigenR_EigenR_isInjective_real', PACKAGE = 'EigenR', M)
+}
+
+EigenR_isInjective_cplx <- function(Re, Im) {
+    .Call('_EigenR_EigenR_isInjective_cplx', PACKAGE = 'EigenR', Re, Im)
+}
+
+EigenR_isSurjective_real <- function(M) {
+    .Call('_EigenR_EigenR_isSurjective_real', PACKAGE = 'EigenR', M)
+}
+
+EigenR_isSurjective_cplx <- function(Re, Im) {
+    .Call('_EigenR_EigenR_isSurjective_cplx', PACKAGE = 'EigenR', Re, Im)
+}
+
+EigenR_isInvertible_real <- function(M) {
+    .Call('_EigenR_EigenR_isInvertible_real', PACKAGE = 'EigenR', M)
+}
+
+EigenR_isInvertible_cplx <- function(Re, Im) {
+    .Call('_EigenR_EigenR_isInvertible_cplx', PACKAGE = 'EigenR', Re, Im)
+}
+
 EigenR_rank_real <- function(M) {
     .Call('_EigenR_EigenR_rank_real', PACKAGE = 'EigenR', M)
 }
@@ -33,6 +65,14 @@ EigenR_inverse_cplx <- function(Re, Im) {
     .Call('_EigenR_EigenR_inverse_cplx', PACKAGE = 'EigenR', Re, Im)
 }
 
+EigenR_pseudoInverse_real <- function(M) {
+    .Call('_EigenR_EigenR_pseudoInverse_real', PACKAGE = 'EigenR', M)
+}
+
+EigenR_pseudoInverse_cplx <- function(Re, Im) {
+    .Call('_EigenR_EigenR_pseudoInverse_cplx', PACKAGE = 'EigenR', Re, Im)
+}
+
 EigenR_kernel_COD_real <- function(M) {
     .Call('_EigenR_EigenR_kernel_COD_real', PACKAGE = 'EigenR', M)
 }
@@ -47,6 +87,14 @@ EigenR_kernel_LU_real <- function(M) {
 
 EigenR_kernel_LU_cplx <- function(Re, Im) {
     .Call('_EigenR_EigenR_kernel_LU_cplx', PACKAGE = 'EigenR', Re, Im)
+}
+
+EigenR_kernelDimension_real <- function(M) {
+    .Call('_EigenR_EigenR_kernelDimension_real', PACKAGE = 'EigenR', M)
+}
+
+EigenR_kernelDimension_cplx <- function(Re, Im) {
+    .Call('_EigenR_EigenR_kernelDimension_cplx', PACKAGE = 'EigenR', Re, Im)
 }
 
 EigenR_image_LU_real <- function(M) {
@@ -111,6 +159,14 @@ EigenR_lsSolve_real <- function(A, b) {
 
 EigenR_lsSolve_cplx <- function(ReA, ImA, Reb, Imb) {
     .Call('_EigenR_EigenR_lsSolve_cplx', PACKAGE = 'EigenR', ReA, ImA, Reb, Imb)
+}
+
+EigenR_lsSolve_cod_real <- function(A, b) {
+    .Call('_EigenR_EigenR_lsSolve_cod_real', PACKAGE = 'EigenR', A, b)
+}
+
+EigenR_lsSolve_cod_cplx <- function(ReA, ImA, Reb, Imb) {
+    .Call('_EigenR_EigenR_lsSolve_cod_cplx', PACKAGE = 'EigenR', ReA, ImA, Reb, Imb)
 }
 
 EigenR_exp_real <- function(M) {