+++ src/EigenR.cpp
@@ -8,13 +8,13 @@
 /* -------------------------------------------------------------------------- */
 Eigen::MatrixXcd matricesToMatrixXcd(const Eigen::MatrixXd& Re,
                                      const Eigen::MatrixXd& Im) {
-  const std::complex<double> I_ {0.0, 1.0};
+  const std::complex<double> I_{0.0, 1.0};
   return Re.cast<std::complex<double>>() + I_ * Im.cast<std::complex<double>>();
 }
 
 Eigen::VectorXcd vectorsToVectorXcd(const Eigen::VectorXd& Re,
                                     const Eigen::VectorXd& Im) {
-  const std::complex<double> I_ {0.0, 1.0};
+  const std::complex<double> I_{0.0, 1.0};
   return Re.cast<std::complex<double>>() + I_ * Im.cast<std::complex<double>>();
 }
 
@@ -54,7 +54,9 @@ Rcpp::ComplexVector cplxMatrixToRcpp(const Eigen::MatrixXcd& M) {
   Rcpp::NumericMatrix outImag(MimagS);
   Rcpp::ComplexMatrix outRealCplx(outReal);
   Rcpp::ComplexMatrix outImagCplx(outImag);
-  Rcomplex I; I.r = 0.0; I.i = 1.0;
+  Rcomplex I;
+  I.r = 0.0;
+  I.i = 1.0;
   Rcpp::ComplexVector out = outRealCplx + I * outImagCplx;
   out.attr("dim") = Rcpp::Dimension(M.rows(), M.cols());
   return out;
@@ -147,6 +149,82 @@ std::complex<double> EigenR_det_sparse_cplx(
   return determinant_sparse<std::complex<double>>(M);
 }
 
+// [[Rcpp::export]]
+double EigenR_absdet(const Eigen::MatrixXd& M) {
+  Eigen::CompleteOrthogonalDecomposition<Eigen::MatrixXd> cod(M);
+  return cod.absDeterminant();
+}
+
+// [[Rcpp::export]]
+double EigenR_logabsdet(const Eigen::MatrixXd& M) {
+  Eigen::CompleteOrthogonalDecomposition<Eigen::MatrixXd> cod(M);
+  return cod.logAbsDeterminant();
+}
+
+/* injective, surjective, invertible ---------------------------------------- */
+template <typename Number>
+bool isInjective(
+    const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>& M) {
+  Eigen::CompleteOrthogonalDecomposition<
+      Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>>
+      cod(M);
+  return cod.isInjective();
+}
+
+// [[Rcpp::export]]
+bool EigenR_isInjective_real(const Eigen::MatrixXd& M) {
+  return isInjective<double>(M);
+}
+
+// [[Rcpp::export]]
+bool EigenR_isInjective_cplx(const Eigen::MatrixXd& Re,
+                             const Eigen::MatrixXd& Im) {
+  const Eigen::MatrixXcd M = matricesToMatrixXcd(Re, Im);
+  return isInjective<std::complex<double>>(M);
+}
+
+template <typename Number>
+bool isSurjective(
+    const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>& M) {
+  Eigen::CompleteOrthogonalDecomposition<
+      Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>>
+      cod(M);
+  return cod.isSurjective();
+}
+
+// [[Rcpp::export]]
+bool EigenR_isSurjective_real(const Eigen::MatrixXd& M) {
+  return isSurjective<double>(M);
+}
+
+// [[Rcpp::export]]
+bool EigenR_isSurjective_cplx(const Eigen::MatrixXd& Re,
+                              const Eigen::MatrixXd& Im) {
+  const Eigen::MatrixXcd M = matricesToMatrixXcd(Re, Im);
+  return isSurjective<std::complex<double>>(M);
+}
+
+template <typename Number>
+bool isInvertible(
+    const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>& M) {
+  Eigen::CompleteOrthogonalDecomposition<
+      Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>>
+      cod(M);
+  return cod.isInvertible();
+}
+
+// [[Rcpp::export]]
+bool EigenR_isInvertible_real(const Eigen::MatrixXd& M) {
+  return isInvertible<double>(M);
+}
+
+// [[Rcpp::export]]
+bool EigenR_isInvertible_cplx(const Eigen::MatrixXd& Re,
+                              const Eigen::MatrixXd& Im) {
+  const Eigen::MatrixXcd M = matricesToMatrixXcd(Re, Im);
+  return isInvertible<std::complex<double>>(M);
+}
+
 /* rank --------------------------------------------------------------------- */
 template <typename Number>
 unsigned rank(const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>& M) {
@@ -169,8 +247,8 @@ unsigned EigenR_rank_cplx(const Eigen::MatrixXd& Re,
 template <typename Number>
 Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic> inverse(
     const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>& M) {
-  const Eigen::FullPivLU<Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>> 
-    lu(M);
+  const Eigen::FullPivLU<Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>>
+      lu(M);
   if(lu.isInvertible()) {
     return lu.inverse();
   } else {
@@ -189,6 +267,29 @@ Rcpp::List EigenR_inverse_cplx(const Eigen::MatrixXd& Re,
   const Eigen::MatrixXcd M = matricesToMatrixXcd(Re, Im);
   const Eigen::MatrixXcd Minv = inverse<std::complex<double>>(M);
   return cplxMatrixToList(Minv);
+}
+
+/* pseudo-inverse ----------------------------------------------------------- */
+template <typename Number>
+Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic> pseudoInverse(
+    const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>& M) {
+  Eigen::CompleteOrthogonalDecomposition<
+      Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>>
+      cod(M);
+  return cod.pseudoInverse();
+}
+
+// [[Rcpp::export]]
+Eigen::MatrixXd EigenR_pseudoInverse_real(const Eigen::MatrixXd& M) {
+  return pseudoInverse<double>(M);
+}
+
+// [[Rcpp::export]]
+Rcpp::List EigenR_pseudoInverse_cplx(const Eigen::MatrixXd& Re,
+                                     const Eigen::MatrixXd& Im) {
+  const Eigen::MatrixXcd M = matricesToMatrixXcd(Re, Im);
+  const Eigen::MatrixXcd pinvM = pseudoInverse<std::complex<double>>(M);
+  return cplxMatrixToList(pinvM);
 }
 
 /* kernel COD --------------------------------------------------------------- */
@@ -241,6 +342,28 @@ Rcpp::List EigenR_kernel_LU_cplx(const Eigen::MatrixXd& Re,
   const Eigen::MatrixXcd M = matricesToMatrixXcd(Re, Im);
   const Eigen::MatrixXcd Kernel = kernel_LU<std::complex<double>>(M);
   return cplxMatrixToList(Kernel);
+}
+
+/* kernel dimension --------------------------------------------------------- */
+template <typename Number>
+unsigned kernelDimension(
+    const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>& M) {
+  Eigen::CompleteOrthogonalDecomposition<
+      Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>>
+      cod(M);
+  return cod.dimensionOfKernel();
+}
+
+// [[Rcpp::export]]
+unsigned EigenR_kernelDimension_real(const Eigen::MatrixXd& M) {
+  return kernelDimension<double>(M);
+}
+
+// [[Rcpp::export]]
+unsigned EigenR_kernelDimension_cplx(const Eigen::MatrixXd& Re,
+                                     const Eigen::MatrixXd& Im) {
+  const Eigen::MatrixXcd M = matricesToMatrixXcd(Re, Im);
+  return kernelDimension<std::complex<double>>(M);
 }
 
 /* image LU ----------------------------------------------------------------- */
@@ -360,8 +483,8 @@ Cholesky<Number> chol(
   if(lltOfM.info() != Eigen::Success) {
     throw Rcpp::exception("The matrix is not positive definite.");
   }
-  const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic> 
-    U = lltOfM.matrixU();
+  const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic> U =
+      lltOfM.matrixU();
   Cholesky<Number> out;
   out.U = U;
   out.determinant = pow(U.diagonal().prod(), 2);
@@ -603,7 +726,7 @@ Rcpp::List EigenR_UtDU_sparse_cplx(const std::vector<size_t>& i,
 }
 */
 
-/* Least-squares ------------------------------------------------------------ */
+/* Least-squares SVD -------------------------------------------------------- */
 template <typename Number>
 Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic> lsSolve(
     const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>& A,
@@ -628,6 +751,34 @@ Rcpp::List EigenR_lsSolve_cplx(const Eigen::MatrixXd& ReA,
   return cplxMatrixToList(X);
 }
 
+/* Least-squares COD -------------------------------------------------------- */
+template <typename Number>
+Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic> lsSolve_cod(
+    const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>& A,
+    const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>& b) {
+  Eigen::CompleteOrthogonalDecomposition<
+      Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>>
+      cod(A);
+  return cod.solve(b);
+}
+
+// [[Rcpp::export]]
+Eigen::MatrixXd EigenR_lsSolve_cod_real(const Eigen::MatrixXd& A,
+                                        const Eigen::MatrixXd& b) {
+  return lsSolve_cod<double>(A, b);
+}
+
+// [[Rcpp::export]]
+Rcpp::List EigenR_lsSolve_cod_cplx(const Eigen::MatrixXd& ReA,
+                                   const Eigen::MatrixXd& ImA,
+                                   const Eigen::MatrixXd& Reb,
+                                   const Eigen::MatrixXd& Imb) {
+  const Eigen::MatrixXcd A = matricesToMatrixXcd(ReA, ImA);
+  const Eigen::MatrixXcd b = matricesToMatrixXcd(Reb, Imb);
+  const Eigen::MatrixXcd X = lsSolve_cod<std::complex<double>>(A, b);
+  return cplxMatrixToList(X);
+}
+
 /* exponential -------------------------------------------------------------- */
 template <typename Number>
 Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic> expm(
@@ -642,7 +793,7 @@ Eigen::MatrixXd EigenR_exp_real(const Eigen::MatrixXd& M) {
 
 // [[Rcpp::export]]
 Rcpp::List EigenR_exp_cplx(const Eigen::MatrixXd& Re,
-                               const Eigen::MatrixXd& Im) {
+                           const Eigen::MatrixXd& Im) {
   const Eigen::MatrixXcd M = matricesToMatrixXcd(Re, Im);
   const Eigen::MatrixXcd Mexp = expm<std::complex<double>>(M);
   return cplxMatrixToList(Mexp);
@@ -722,7 +873,7 @@ Eigen::MatrixXd EigenR_cosh_real(const Eigen::MatrixXd& M) {
 
 // [[Rcpp::export]]
 Rcpp::List EigenR_cosh_cplx(const Eigen::MatrixXd& Re,
-                           const Eigen::MatrixXd& Im) {
+                            const Eigen::MatrixXd& Im) {
   const Eigen::MatrixXcd M = matricesToMatrixXcd(Re, Im);
   const Eigen::MatrixXcd Mcosh = coshm<std::complex<double>>(M);
   return cplxMatrixToList(Mcosh);
@@ -751,7 +902,7 @@ Rcpp::List EigenR_sinh_cplx(const Eigen::MatrixXd& Re,
 /* power -------------------------------------------------------------------- */
 template <typename Number>
 Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic> powm(
-    const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>& M, 
+    const Eigen::Matrix<Number, Eigen::Dynamic, Eigen::Dynamic>& M,
     const Number& p) {
   return M.pow(p);
 }
@@ -763,8 +914,8 @@ Eigen::MatrixXd EigenR_pow_real(const Eigen::MatrixXd& M, const double& p) {
 
 // [[Rcpp::export]]
 Rcpp::List EigenR_pow_cplx(const Eigen::MatrixXd& Re,
-                            const Eigen::MatrixXd& Im,
-                            const std::complex<double>& p) {
+                           const Eigen::MatrixXd& Im,
+                           const std::complex<double>& p) {
   const Eigen::MatrixXcd M = matricesToMatrixXcd(Re, Im);
   const Eigen::MatrixXcd Mpow = powm<std::complex<double>>(M, p);
   return cplxMatrixToList(Mpow);
@@ -789,4 +940,3 @@ Rcpp::List EigenR_sqrt_cplx(const Eigen::MatrixXd& Re,
   const Eigen::MatrixXcd Msqrt = sqrtm<std::complex<double>>(M);
   return cplxMatrixToList(Msqrt);
 }
-