+++ src/RcppExports.cpp
@@ -64,6 +64,97 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// EigenR_absdet
+double EigenR_absdet(const Eigen::MatrixXd& M);
+RcppExport SEXP _EigenR_EigenR_absdet(SEXP MSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type M(MSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_absdet(M));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EigenR_logabsdet
+double EigenR_logabsdet(const Eigen::MatrixXd& M);
+RcppExport SEXP _EigenR_EigenR_logabsdet(SEXP MSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type M(MSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_logabsdet(M));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EigenR_isInjective_real
+bool EigenR_isInjective_real(const Eigen::MatrixXd& M);
+RcppExport SEXP _EigenR_EigenR_isInjective_real(SEXP MSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type M(MSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_isInjective_real(M));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EigenR_isInjective_cplx
+bool EigenR_isInjective_cplx(const Eigen::MatrixXd& Re, const Eigen::MatrixXd& Im);
+RcppExport SEXP _EigenR_EigenR_isInjective_cplx(SEXP ReSEXP, SEXP ImSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Re(ReSEXP);
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Im(ImSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_isInjective_cplx(Re, Im));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EigenR_isSurjective_real
+bool EigenR_isSurjective_real(const Eigen::MatrixXd& M);
+RcppExport SEXP _EigenR_EigenR_isSurjective_real(SEXP MSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type M(MSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_isSurjective_real(M));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EigenR_isSurjective_cplx
+bool EigenR_isSurjective_cplx(const Eigen::MatrixXd& Re, const Eigen::MatrixXd& Im);
+RcppExport SEXP _EigenR_EigenR_isSurjective_cplx(SEXP ReSEXP, SEXP ImSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Re(ReSEXP);
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Im(ImSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_isSurjective_cplx(Re, Im));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EigenR_isInvertible_real
+bool EigenR_isInvertible_real(const Eigen::MatrixXd& M);
+RcppExport SEXP _EigenR_EigenR_isInvertible_real(SEXP MSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type M(MSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_isInvertible_real(M));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EigenR_isInvertible_cplx
+bool EigenR_isInvertible_cplx(const Eigen::MatrixXd& Re, const Eigen::MatrixXd& Im);
+RcppExport SEXP _EigenR_EigenR_isInvertible_cplx(SEXP ReSEXP, SEXP ImSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Re(ReSEXP);
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Im(ImSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_isInvertible_cplx(Re, Im));
+    return rcpp_result_gen;
+END_RCPP
+}
 // EigenR_rank_real
 unsigned EigenR_rank_real(const Eigen::MatrixXd& M);
 RcppExport SEXP _EigenR_EigenR_rank_real(SEXP MSEXP) {
@@ -110,6 +201,29 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// EigenR_pseudoInverse_real
+Eigen::MatrixXd EigenR_pseudoInverse_real(const Eigen::MatrixXd& M);
+RcppExport SEXP _EigenR_EigenR_pseudoInverse_real(SEXP MSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type M(MSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_pseudoInverse_real(M));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EigenR_pseudoInverse_cplx
+Rcpp::List EigenR_pseudoInverse_cplx(const Eigen::MatrixXd& Re, const Eigen::MatrixXd& Im);
+RcppExport SEXP _EigenR_EigenR_pseudoInverse_cplx(SEXP ReSEXP, SEXP ImSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Re(ReSEXP);
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Im(ImSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_pseudoInverse_cplx(Re, Im));
+    return rcpp_result_gen;
+END_RCPP
+}
 // EigenR_kernel_COD_real
 Eigen::MatrixXd EigenR_kernel_COD_real(const Eigen::MatrixXd& M);
 RcppExport SEXP _EigenR_EigenR_kernel_COD_real(SEXP MSEXP) {
@@ -153,6 +267,29 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Re(ReSEXP);
     Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Im(ImSEXP);
     rcpp_result_gen = Rcpp::wrap(EigenR_kernel_LU_cplx(Re, Im));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EigenR_kernelDimension_real
+unsigned EigenR_kernelDimension_real(const Eigen::MatrixXd& M);
+RcppExport SEXP _EigenR_EigenR_kernelDimension_real(SEXP MSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type M(MSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_kernelDimension_real(M));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EigenR_kernelDimension_cplx
+unsigned EigenR_kernelDimension_cplx(const Eigen::MatrixXd& Re, const Eigen::MatrixXd& Im);
+RcppExport SEXP _EigenR_EigenR_kernelDimension_cplx(SEXP ReSEXP, SEXP ImSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Re(ReSEXP);
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Im(ImSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_kernelDimension_cplx(Re, Im));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -350,6 +487,32 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// EigenR_lsSolve_cod_real
+Eigen::MatrixXd EigenR_lsSolve_cod_real(const Eigen::MatrixXd& A, const Eigen::MatrixXd& b);
+RcppExport SEXP _EigenR_EigenR_lsSolve_cod_real(SEXP ASEXP, SEXP bSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type A(ASEXP);
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type b(bSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_lsSolve_cod_real(A, b));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EigenR_lsSolve_cod_cplx
+Rcpp::List EigenR_lsSolve_cod_cplx(const Eigen::MatrixXd& ReA, const Eigen::MatrixXd& ImA, const Eigen::MatrixXd& Reb, const Eigen::MatrixXd& Imb);
+RcppExport SEXP _EigenR_EigenR_lsSolve_cod_cplx(SEXP ReASEXP, SEXP ImASEXP, SEXP RebSEXP, SEXP ImbSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type ReA(ReASEXP);
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type ImA(ImASEXP);
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Reb(RebSEXP);
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Imb(ImbSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_lsSolve_cod_cplx(ReA, ImA, Reb, Imb));
+    return rcpp_result_gen;
+END_RCPP
+}
 // EigenR_exp_real
 Eigen::MatrixXd EigenR_exp_real(const Eigen::MatrixXd& M);
 RcppExport SEXP _EigenR_EigenR_exp_real(SEXP MSEXP) {
@@ -542,14 +705,26 @@ static const R_CallMethodDef CallEntries[] = {
     {"_EigenR_EigenR_det_cplx", (DL_FUNC) &_EigenR_EigenR_det_cplx, 2},
     {"_EigenR_EigenR_det_sparse_real", (DL_FUNC) &_EigenR_EigenR_det_sparse_real, 5},
     {"_EigenR_EigenR_det_sparse_cplx", (DL_FUNC) &_EigenR_EigenR_det_sparse_cplx, 5},
+    {"_EigenR_EigenR_absdet", (DL_FUNC) &_EigenR_EigenR_absdet, 1},
+    {"_EigenR_EigenR_logabsdet", (DL_FUNC) &_EigenR_EigenR_logabsdet, 1},
+    {"_EigenR_EigenR_isInjective_real", (DL_FUNC) &_EigenR_EigenR_isInjective_real, 1},
+    {"_EigenR_EigenR_isInjective_cplx", (DL_FUNC) &_EigenR_EigenR_isInjective_cplx, 2},
+    {"_EigenR_EigenR_isSurjective_real", (DL_FUNC) &_EigenR_EigenR_isSurjective_real, 1},
+    {"_EigenR_EigenR_isSurjective_cplx", (DL_FUNC) &_EigenR_EigenR_isSurjective_cplx, 2},
+    {"_EigenR_EigenR_isInvertible_real", (DL_FUNC) &_EigenR_EigenR_isInvertible_real, 1},
+    {"_EigenR_EigenR_isInvertible_cplx", (DL_FUNC) &_EigenR_EigenR_isInvertible_cplx, 2},
     {"_EigenR_EigenR_rank_real", (DL_FUNC) &_EigenR_EigenR_rank_real, 1},
     {"_EigenR_EigenR_rank_cplx", (DL_FUNC) &_EigenR_EigenR_rank_cplx, 2},
     {"_EigenR_EigenR_inverse_real", (DL_FUNC) &_EigenR_EigenR_inverse_real, 1},
     {"_EigenR_EigenR_inverse_cplx", (DL_FUNC) &_EigenR_EigenR_inverse_cplx, 2},
+    {"_EigenR_EigenR_pseudoInverse_real", (DL_FUNC) &_EigenR_EigenR_pseudoInverse_real, 1},
+    {"_EigenR_EigenR_pseudoInverse_cplx", (DL_FUNC) &_EigenR_EigenR_pseudoInverse_cplx, 2},
     {"_EigenR_EigenR_kernel_COD_real", (DL_FUNC) &_EigenR_EigenR_kernel_COD_real, 1},
     {"_EigenR_EigenR_kernel_COD_cplx", (DL_FUNC) &_EigenR_EigenR_kernel_COD_cplx, 2},
     {"_EigenR_EigenR_kernel_LU_real", (DL_FUNC) &_EigenR_EigenR_kernel_LU_real, 1},
     {"_EigenR_EigenR_kernel_LU_cplx", (DL_FUNC) &_EigenR_EigenR_kernel_LU_cplx, 2},
+    {"_EigenR_EigenR_kernelDimension_real", (DL_FUNC) &_EigenR_EigenR_kernelDimension_real, 1},
+    {"_EigenR_EigenR_kernelDimension_cplx", (DL_FUNC) &_EigenR_EigenR_kernelDimension_cplx, 2},
     {"_EigenR_EigenR_image_LU_real", (DL_FUNC) &_EigenR_EigenR_image_LU_real, 1},
     {"_EigenR_EigenR_image_LU_cplx", (DL_FUNC) &_EigenR_EigenR_image_LU_cplx, 2},
     {"_EigenR_EigenR_image_QR_real", (DL_FUNC) &_EigenR_EigenR_image_QR_real, 1},
@@ -566,6 +741,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_EigenR_EigenR_UtDU_cplx", (DL_FUNC) &_EigenR_EigenR_UtDU_cplx, 2},
     {"_EigenR_EigenR_lsSolve_real", (DL_FUNC) &_EigenR_EigenR_lsSolve_real, 2},
     {"_EigenR_EigenR_lsSolve_cplx", (DL_FUNC) &_EigenR_EigenR_lsSolve_cplx, 4},
+    {"_EigenR_EigenR_lsSolve_cod_real", (DL_FUNC) &_EigenR_EigenR_lsSolve_cod_real, 2},
+    {"_EigenR_EigenR_lsSolve_cod_cplx", (DL_FUNC) &_EigenR_EigenR_lsSolve_cod_cplx, 4},
     {"_EigenR_EigenR_exp_real", (DL_FUNC) &_EigenR_EigenR_exp_real, 1},
     {"_EigenR_EigenR_exp_cplx", (DL_FUNC) &_EigenR_EigenR_exp_cplx, 2},
     {"_EigenR_EigenR_log_real", (DL_FUNC) &_EigenR_EigenR_log_real, 1},