+++ NEWS.md
@@ -1,3 +1,23 @@
+# EigenR 1.2.0
+
+New functions:
+
+- `Eigen_pinverse` returns the pseudo-inverse (Moore-Penrose generalized 
+inverse) of a matrix.
+
+- `Eigen_absdet` returns the absolute value of the determinant of a real square 
+matrix.
+
+- `Eigen_logabsdet` returns the logarithm of the absolute value of the 
+determinant of a real square matrix.
+
+- `Eigen_isInjective`, `Eigen_isSurjective`, `Eigen_isInvertible` return a Boolean value.
+
+- `Eigen_kernelDimension` returns the dimension of the kernel of a matrix.
+
+- New method for `Eigen_lsSolve`, faster.
+
+
 # EigenR 1.1.0
 
 New matricial functions have been included: exponential, logarithm, cosine,